Optimising update broadcast strategy
Previously, we broadcast update information for the general and session channels in a piecemeal fashion, causing us to have to loop through all clients for each piece.
Now, we consolidate all updates into a single hash for each channel, then just broadcast that to all relevant clients for that channel.