Implementing a watcher that periodically checks if WebSockets is alive, and restarts it if it's dead.
The keepAlive watcher tries to send a ping, eg. the character 'p', through the WebSocket connection every KEEP_ALIVE_INTERVAL.
If the .send() doesn't succeed, it assumes that the WebSocket is dead, and re-initialises it. After another KEEP_ALIVE_INTERVAL, it attempts the ping again, re-initialising if it fails.
This ping/re-initialise process repeats till a successful ping is sent.
In the worst case, if the server/connection dies, the watcher _should_ be able to re-establish a working WebSocket within 2 x KEEP_ALIVE_INTERVAL of the server/connection becoming responsive again.