Switching to use lists of lists to broadcasts session stats instead of a dictionary.
Now a session broadcast looks like this: [['session1', 123], ['session2', 99], ['session3', 33]]
instead of: {"session1": 123, "session3":33, "session2":99}

The benefit of using lists is that sessions can now be sorted() by number of votes (since dictionaries can't be sorted).
However, this means our code is slightly less understandable as we now need to address values by numeric index, eg. session[0] for the session name and session[1] for number of votes
We could have the sessions list as a list of dictionaries. That would still work with sorted(), but it would mean far more bandwidth overhead in sending those extra characters.